Update to note case sensitivity of bundle IDs
Based on some recent feedback from field working with a customer, it was found that bundle IDs are case sensitive, and policies will not apply properly if bundle IDs are not entered with the right casing. Added a couple of sentences to this article to note this - please feel free to edit the wording as needed. Thank you!