Remove Wallet as example from hide apps section
Through some feedback from the field working with customers, it was found that the Wallet app CAN be hidden on iOS devices. Our docs stated that the Wallet app cannot be hidden as an example - this update removes that example and keeps the Settings app as the only example of apps that cannot be hidden.